# docs/swagger.yaml
basePath: /api/v1
definitions:
  entities.AdminLoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  entities.AdminLoginResponse:
    properties:
      role:
        type: string
      token:
        type: string
    type: object
...
  entities.DetectColumnsResponse:
    properties:
      columns:
        items:
          type: string
        type: array
    type: object
...
  entities.ImportError:
    properties:
      field:
        type: string
      message:
        type: string
      row:
        type: integer
    type: object
...
  entities.ImportSummary:
    properties:
      errors:
        items:
          $ref: '#/definitions/entities.ImportError'
        type: array
      new_products:
        type: integer
      skipped_rows:
        type: integer
      total_rows:
        type: integer
      updated_products:
        type: integer
    type: object
...
  entities.ProviderMapPin:
    properties:
      business_name:
        type: string
      city:
        type: string
      created_at:
        type: string
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      state:
        type: string
    type: object
...
  entities.VerifyOTPRequest:
    properties:
      code:
        type: string
    required:
    - code
    type: object
  entities.VerifyOTPResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  responses.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: soporte@visionprice.app
    name: VisionPrice Support
  description: API backend para el módulo de Proveedores de VisionPrice
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: VisionPrice Proveedores API
  version: "1.0"
paths:
  /api/v1/admin/geography/providers:
    get:
      description: Retorna ubicaciones de proveedores activos con datos seguros (sin
        RFC, email, precios)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entities.ProviderMapPin'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Obtener pines de proveedores para el mapa (HU_SYS_02)
      tags:
      - Admin
  /api/v1/admin/login:
    post:
      consumes:
      - application/json
      description: Autentica un usuario administrador y devuelve un JWT con rol USER_SYS_ADMIN
      parameters:
      - description: Credenciales de admin
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.AdminLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/entities.AdminLoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Login de administrador del sistema
      tags:
      - Admin
...
  /api/v1/auth/2fa/generate:
    post:
      consumes:
      - application/json
      description: Genera y envía un código OTP de 6 dígitos al proveedor (requiere
        token temporal de login)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Generar código OTP
      tags:
      - 2FA
  /api/v1/auth/2fa/verify:
    post:
      consumes:
      - application/json
      description: Valida el código OTP y emite tokens de acceso y refresh
      parameters:
      - description: Código OTP
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.VerifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/entities.VerifyOTPResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Verificar código OTP
      tags:
      - 2FA
...
  /api/v1/extractions/detect-columns:
    post:
      consumes:
      - multipart/form-data
      description: Sube un archivo Excel y retorna las columnas detectadas en los
        headers
      parameters:
      - description: Archivo Excel (.xlsx)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/entities.DetectColumnsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Detectar columnas de archivo Excel
      tags:
      - Extracciones
  /api/v1/extractions/import:
    post:
      consumes:
      - multipart/form-data
      description: Procesa un archivo Excel usando el mapeo guardado y crea/actualiza
        productos
      parameters:
      - description: Archivo Excel (.xlsx)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/entities.ImportSummary'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Importar productos desde Excel
      tags:
      - Extracciones
...
securityDefinitions:
  BearerAuth:
    description: 'Ingrese el token JWT con el prefijo Bearer. Ejemplo: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"